requirements.txt
requests
python-nmap
reportlab
colorama
beautifulsoup4
rich
flask
